--- project/ms.supp
@@ -1,12 +1,0 @@
-{
-   ignore_libreadline_memory_errors
-   Memcheck:Leak
-   ...
-   obj:*/libreadline.so.*
-}
-
-{
-   ignore_fork_memory
-   Memcheck:Leak
-   fun:fork
-}
