--- project/valcodes.txt
@@ -1,1 +1,0 @@
-valgrind --leak-check=full --show-leak-kinds=all --suppressions=ms.supp --log-file=leakreport.txt ./minishell
